Langfuse/requirements.txt
flask
langfuse==2.23.1
openai==0.28
python-dotenv